fix(core): harden stdio lifecycle concurrency and prevent token total overflow
- StdioTransport: synchronize StartAsync/DisposeAsync state transitions to avoid races
- StdioTransport: prevent stale reader task from overriding active connection state
- Telemetry: migrate token aggregate totals from int to long
- Evaluation: use long for TotalTokens and validate percentile input range

+++ src/Dawning.Agents.Abstractions/Telemetry/ITokenUsageTracker.cs
@@ -67,17 +67,17 @@ public interface ITokenUsageTracker
     /// <summary>
     /// 总输入 Token 数
     /// </summary>
-    int TotalPromptTokens { get; }
+    long TotalPromptTokens { get; }
 
     /// <summary>
     /// 总输出 Token 数
     /// </summary>
-    int TotalCompletionTokens { get; }
+    long TotalCompletionTokens { get; }
 
     /// <summary>
     /// 总 Token 数
     /// </summary>
-    int TotalTokens { get; }
+    long TotalTokens { get; }
 
     /// <summary>
     /// 总调用次数
